requirements.txt
bark @ git+https://github.com/suno-ai/bark.git
py-cord>=2.3.1
langchain>=0.0.262
openai_whisper>=20230314
python-dotenv>=1.0.0
scipy>=1.11.1
torch>=2.0.1
torchaudio>=2.0.2
tortoise-tts @ git+https://github.com/neonbjb/tortoise-tts.git
ffmpeg-python >= 0.2.0
